List command the script is blocking on.
While running the verify or test script, the scripts will block with a
message like this:

  - Waiting for task completion (up to 1200 seconds)

This is while waiting for a long running task, like provisioning a
machine.  This trivial change outputs the command being run by the
function, to give a hint about what it's blocking on.

The command being run is usually another shell function with some
arguments, but so far it gives a pretty good hint about what's going
on.  For example:

  - Waiting for task completion (up to 1200 seconds)  - Command:
    'check_provisioning_status centos-0 master-0 provisioned'