Use clouds.yaml file and yq to merge with existing user's config
This changes the approach to use a clouds.yaml file in the repo, and
uses yq to merge with a user's existing clouds.yaml. This has the
benefit of overriding the existing "metal3" config if it is ever
updated.

We are also renaming the clouds.yaml configurations in OpenShift's
dev-scripts, which previously used the metal3 name.