Remove RDO requirement for RHEL8
This may also enable testing on CentOS8 where the RDO repos installed
for CentOS8 do not yet exist.

The dependencies on the RDO repos were virtualbmc (now containerized)
newer qemu-kvm (now in RHEL/CentOS8) and ironicclient/openstackclient.

The remaining client dependency is moved to pip for now, but we should
probably look at running that from a container image instead.