Make installing openstack/ironic clients optional
For RHEL/CentOS this has proven problematic because the RDO repos
aren't always available for a specific release (e.g RHEL7 and
recently CentOS 8 took a while to become available)

We could rely on pip for all platforms, but this can cause
dependency issues with packaged python packages, and typically
can't be used for debugging on a target cluster (whereas a
container image often can).

So this change provides a shell alias and a wrapper script to provide
a backwards compatible interface, but launch the new ironic-client
image instead - testing indicates this is slightly slower but it
should be sufficient for debug usage which is the primary purpose
of the clients in this context.