Allow customization of hostnames for nodes
The master-X hostname format came out of an OpenShift requirement
for etcd bootstrapping. We're making changes to eliminate that
requirement and it is desirable to be able to test those changes in
a virtual environment such as metal3-dev-env. This change allows
arbitrary formats for the hostnames of the defined flavors.

The way it works is that the user sets an Ansible variable named
{{flavor}}_hostname_format. This is a format string with exactly one
placeholder. The placeholder will be formatted with the number of
the node and that will become the node's hostname.

For example, if the master_hostname_format is set to "foo%sbar", the
resulting hostnames will be foo0bar, foo1bar, etc.