Export CONTAINER_RUNTIME so subshells can use it
The container runtime used in the dev-env is decided by the environment variable CONTAINER_RUNTIME. By default, its value is Docker when the host is Ubuntu, and it is Podman when the host is Centos. This variable is set in lib/common.sh, and every script that sources this file has the right value of this variable. However, the dev-env also calls some scripts from other repos (e.g. Baremetal operator) that don't source the file, and they use the wrong container runtime. This leads to that when we run the dev-env on Ubuntu where Podman is not installed, but it tries to use Podman, resulting in an error.