CI: backup and restore libvirt VM
To speed up feature testing, VMs are backed up when
an appropriate cluster setup created.
After a test is run the base setup can be restored.

Cluster deprovisioning, which is time consuming,
is no longer needed after every test case
because the same state can be achieved with the restore.