Add support for VLAN in integration tests
This change allows setting the `EXTERNAL_VLAN_ID` environment variable
in metal3-dev-env. When this is set, the target cluster's bare metal
network is configured to use this VLAN. When this variable is not set,
all existing behaviour is preserved.

Currently, the VM nodes have two interfaces, one used for the
provisioning network, the other used for the bare metal network. This
change allows for a setup where target hosts have a single interface
with link and a dedicated VLAN for the bare metal network.