Add renovate config with Minikube updates
Add a renovate configuration which will automatically raise pull
requests for minikube updates.

Renovate pull requests will be automatically assigned to two of the
members of the OWNERS file who support the CI for this project. We could
automate the assignees list by using a CODEOWNERS file in future.

If this change is successful, we should see Renovate try to bump
minikube to version v1.21.0.