Verify existence of resources before filtering
For various reasons it may happen that communication with the k8s API
fails. When attempting to filter this empty result, it naturally causes
errors that can fail the whole task. To avoid this we can check that
the module succeeded in communicating with k8s. Now if the module
fails for some reason it can retry instead of stopping due to an error
when trying to filter.