Check NUM_NODES equal to control plane + workers
We only set NUM_OF_MASTER_REPLICAS=3 and
NUM_OF_WORKER_REPLICAS=1 for upgrade tests, but NUM_NODES
defaults to 2 unless overridden. This moves the variables to
upgrade_vars.sh from the upgrade Makefile and checks that the variables
make sence in common.sh. The reason for moving the variables is that
upgrade_vars.sh is sourced before creating the VMs, whereas the Makefile
is executed after this.