network: allow use of NATv6 if needed
Since Libvirt 6.5, NATv6 works with IPv6. If a user requests an IPv6
network today with metal3-dev-env, NAT won't work as you have to opt-in
to the feature. This adds the required `ipv6=yes` to the `<nat>` tag to
make this work, if a user sets `enable_natv6` on that network.

On previous versions of libvirt the unknown attribute just gets ignored
and removed.