Use kubectl instead of k8s module for applying
The k8s module does no validation on the yaml when applying! It's like
running kubectl with --validate=false. Furthermore, even adding the
validate block to the module won't validate custom resources. This is
quite unexpected, and meant that we were applying invalid templates
(nodeDrainTimeout was skipped from our definitions.)