Add quay.io as the registry
Because of recent change regarding container registry parameterization a bug  was introduced in metal3-dev-env which resulted in the local images not being checked in case of PRs in CAPM3, BMO, IPAM etc repositories. The reason being the se commands are trying to find registry.nordix.org/quay-io-proxy/metal3-io/baremetal-operator as image in manager.yml but in manager.yaml of different repositories we have quay.io. This resulted in sed being a no-op and local images were not tested anymore. This quick fix is falling back to quay.io until proper fix lands.