Don't start fetching CRs until m3Machine is rebuilt after pivot back
We call fetch_manifests.sh after we pivot back and it starts very
quick even before Metal3Machine controller rebuilds the status
after metal3Machine objects are pivoted back. This patch holds on
running fetch_manifests.sh script until the status of metal3Machine
is available.
Note: when clusterctl move, status is not moved, and it is job of
controllers to rebuild the status.