CI: prepare for the new ironic container
We will have a single ironic executable instead of API+conductor,
and the default setup will not have mariadb. HTTPd will be used for both
serving images and terminating TLS.

Some ignore_errors had to be temporary inserted, so that the job does
not fail on non-existing containers. To be reverted once the migration
is done.