Handle zero length dhcp_range variables
In this case, we treat it the same as the lists being undefined since
it implies DHCP is disabled.

This means it's possible to do a conditional in a vars file e.g:

dhcp_range_v4: "{{ [ <IP>, <IP>] if lookup('env', 'STATIC_IPS') == '' else [] }}"

Which is more convenient that conditionally generating the map keys.