Allow overriding CAPM3_VERSION and CAPI_VERSION in upgrade vars
Also remove hard coded releases for CAPI and CAPM3. The latest release
for the chosen API version will be set automatically.

The idea is to make it possible to test multiple scenarios with the
upgrade test. For example, upgrading just a patch (v1.0.0 -> v1.0.1).
The default is to upgrade from the latest CAPI v1alpha4 release to the
latest v1beta1 release.