Ensure that machines are provisioned and running before repivot
This check was missing in upgrade before repivot.
Since we are doing this check in multiple places, I decided to factor
out the steps into a separate file that can be included where needed.
When doing this we must ensure that the correct kubeconfig is used by
passing it as a variable.