add logic to verify passwordless sudo
This fix is implemented because on a freshly cloned or fully cleaned metal3-dev-env
it is possible to bootstrap a cluster without the passwordless sudo
configuration but at subsequent executions the script will stop running
and one of the components will throw an error to notify the user that it needs
passwordless sudo.

In order to avoid running the dev-env with incorrect sudo configuration this
commit implements a verification step that will stop the execution in case
the sudo configuratio is incorrect.