🐛 Fix BMO and Ironic image updates
The function used to update the container images for BMO and Ironic only
worked for replacing the original image with the one in the local
registry. This fix makes it possible to use a different images and more
importantly tags, so that we can start BMO with a specific tag and then
upgade to latest.

This commit also unifies the image replacement so that it works similar
for CAPM3, BMO and Ironic.