Make IMAGE_OS value lower case
IMAGE_OS used to take values like "Ubuntu" or "Centos". This commit
changes it to "ubuntu" or "centos". This makes the variable useful in
places were it would be awkward to have a capitalized value. It will
allow us to remove some logic and extra variables from CI since we can
use IMAGE_OS directly.

Note that both capitalized and lower case values will be supported while
we work on changing the CI to use lower case values.