Use the same config for Ironic after pivot
Before pivot, Ironic can run either as a local container on the host or
in the bootstrap cluster. After pivot, it always runs in the workload
cluster. There was an issue when pivoting from a local container into
the workload cluster that resulted in wrong configuration for Ironic.

The issue comes from the fact that we configure the local Ironic through
environment variables directly, but the in-cluster Ironic is configured
through deployment manifests that should be edited based on the
environment variables. What happened was that we never edited these
manifests when going for a local Ironic, which meant that the manifests
didn't have the correct configuration when pivoting.

To solve the issue, I have rearanged the 03 script so that we always
edit the manifests (even when running Ironic locally).

The second issue is that we restore all changes we make to the manifests
in 03_launch_mgmt_cluster.sh before pivoting. The changes are still
needed to get the correct images during pivot, upgrade and re-pivot.
For this reason I have removed the cleanup of these files. They are all
restored the next time 03_launch_mgmt_cluster.sh is run unless
FORCE_REPO_UPDATE is set to false. The dirty repos can also be cleaned
manually with a `git reset --hard`.