improve metal3-dev-env repository cloning
Two issues are being fixed by this commit.

1. Currently the dev-env have 2 independent repo
cloning functionality. One for building local images
and one for customizing CRDs.

The two cloning process is incompatible because the cloning
used for CRD customization happens after the images
specified with _LOCAL_IMAGE postfix have been
built so a clean environment where the repositories
are not yet cloned to $HOME/go/src/github.com/metal3-io/
would not support the local image building based on the same repos
that will be used for CRD customization in later stages of the bootstrap
process.

2. The clone_repo bash function had no support for cloning
specific commits from branches of CAPI, CAPM3 and IPAM.

NOTE: With the changes in this commit the git repo support
implemented exclusively for _LOCAL_IMAGE variables will become redundant
because the _LOCAL_IMAGEs can use the same repositories that are
now cloned by the clone_repo functions thus the redundant functionality
should be removed in later commits in order to keep the scope of this
commit manageable.