improve local build of CAPI,CAPM3,IPAM,BMO
This commit improves 3 different things.

1. Add default build option for CAPI,CAPM3,IPAM,BMO.
If the default build option is set to true for any of the
aforementioned repositories then the dev-env will build
the container images based on the same directory that is used
for CRD customization.

2. In order to unify the local container image build process
with the CRD customization process this commit removes the git
repository supoort feature from 02 script because that was exclusively
wirrten to support the local image building thus duplicating the
repository cloning process and by default separating the directories
used for CRD customization and image building.

3. Fix CAPI build options for local image building.