IPA downloader exit on error and retry
We used to wait for the container to finnish, which just prints the exit
code. If we instead run it without -d, it will block on that step and
propagate the exit code properly.

For retry we also need to add --rm since otherwise it will complain
about already existing container.