bind registry to an EXTERNAL_SUBNET_V4_HOST
Registry should be binding to an IP by default. In case dev-env
is ran on a machine that has internet connectivity, binding to all
interfaces exposes unauthenticated registry to anyone. This would
allow anyone to replace images with arbitrary code, or perform DOS
by filling the machine's disk, etc.

Here we replace binding "$REGISTRY_PORT":5000 with $REGISTRY:5000
as $REGISTRY is defined as $EXTERNAL_SUBNET_V4_HOST:$REGISTRY_PORT
and by default that is 192.168.11.1:5000