unify ansible extra_var handling
Co-authored-by: Tuomo Tanskanen <tuomo.tanskanen@est.tech>
Co-authored-by: @tuminoid

It has been noticed that a firewalld related ansible extra_var was
passed to Ansible in json format instead of regular key=value format.

The root cause of the isse was that this variable was handled as bool type and
directly used in ansible "when" conditions.

I beleive it is a less complex aproach if ansible arguments are all strings
in bash, then if the variable would be used in a condition in ansible,
it can be type casted via a type filter at the place of use thus the cotributors
modifying the bash code have no need to one by one investigate all the
use-cases af all possible extra_var of ansible.

It has been decided to unify the extra_var syntax in the bash scripts
to follow the "key=value" format.

This commit also removes an unnecessary external subnet related
extra_var as it was actually both redundant and acted as a noop.