add mitigation for kubeadm join issue
For some reason kubeadm can't validate the api server because
of a JSW error, but it seems the issue is related to timing.

These additional delays introduced in this commit are not a real solution to
the problem, but something is needed to unclog the CI and decrease the number
of failures experienced on local dev-env deployments.

This commit also removes some keepalived monitoring code fragments that are not needed and
some keepalived example config data that is not needed either.