pin downloaded scripts and binaries to a sha digest
For downloaded install scripts and readily-built binaries, we need to
add a custom function to download them and verify the SHA in one go.
This also means that when doing version update, we need to also change
the SHA.

For testing purposes, we provide an insecure override variable
INSECURE_SKIP_DOWNLOAD_VERIFICATION, which if set to true, will not
verify the downloads.