move sourcing of libs after initial package install
Many of the lib/*.sh actually run code requiring additional command
line tooling, which is only installed after sourcing them. Running
"make" on the vanilla Ubuntu fails currently.

This is kind of first-aid to fix the issue. The real solution would be
for the libs would be not to run any code when sourced, only create
functions for later use.