Allow proxy connection when dividing response
If a proxy is used when trying to connect to the server, a response message
with the content "HTTP/1.1 200 Connection established" will be returned in the
first line, followed by the required header information in the second line,
which will result in incorrect headers and body being obtained. This patch aims
to address this issue.

Signed-off-by: Guangwen Feng <fenggw-fnst@fujitsu.com>