fix KUBECTL_SHA256 to match KUBERNETES_VERSION
When KUBERNETES_VERSION is changed, it changes the KUBECTL_SHA256 as
well as the versions are coupled. There is TODO item to get rid of
this coupling.

The issue is not visible until CI images are rebuilt as 01_script
does not get executed during regular integration tests. Anyone using
vanilla image will get the error though, like Ironic CI.