fix the application of the kvm_tdp_mmu workaround
At the time of writing the CI is experiencing broken pipe issues for a week.
The issues are present for both Centos and Ubuntu.
In the previous implementation the fix was applied in 01 script but that is not executed
in most of the scripts currently. This commit moves the application of the fix into 02 script and
enables it for both distros again.