tie BMO commits for CAPM3RELEASEBRANCH variables
This commit:
  - ties specific commit hashes of the BMO release tags for each
    pre-defined CAPM3RELEASEBRANCH variable in the lib/common.sh

The reason is that in case BMO release branch HEAD is used together
with a specific BMO container image release tag in the same deployment, there
is a chance that a e.g. some patch could introduce CRD misalignment between
the unchanged  container image and the moving release branch head thus causing
BMO operations or even BMO bootstrapping to fail.

Signed-off-by: Adam Rozman <adam.rozman@est.tech>