Wait for local ironic to be up
This ensures that ironic is ready before proceeding when running ironic
locally. It is specifically needed when SKIP_APPLY_BMH is set to true.
In this case we will otherwise continue straight to 04_verify.sh with a
high probability of finding the ironic endpoint unavailable.

When SKIP_APPLY_BMH is NOT true, we will in any case implicitly wait for
ironic, since the BMHs won't become ready until ironic does.

Signed-off-by: Lennart Jern <lennart.jern@est.tech>